+++ d3web-Essentials/d3web-Kernel/src/main/java/de/d3web/core/inference/condition/CondNumGreater.java
@@ -44,7 +44,7 @@ public class CondNumGreater extends CondNum {
 
 	@Override
 	protected boolean compare(Double caseValue, Double conditionedValue) {
-		return (caseValue.doubleValue() > conditionedValue.doubleValue());
+		return (caseValue > conditionedValue);
 	}
 
 	@Override
