+++ d3web-Essentials/d3web-Kernel/src/main/java/de/d3web/core/inference/condition/CondNumLess.java
@@ -45,7 +45,7 @@ public class CondNumLess extends CondNum {
 
 	@Override
 	protected boolean compare(Double caseValue, Double conditionedValue) {
-		return (caseValue.doubleValue() < conditionedValue.doubleValue());
+		return (caseValue < conditionedValue);
 	}
 
 	@Override
