+++ d3web-Essentials/d3web-Kernel/src/main/java/de/d3web/core/inference/condition/CondNumLessEqual.java
@@ -43,7 +43,7 @@ public class CondNumLessEqual extends CondNum {
 
 	@Override
 	protected boolean compare(Double caseValue, Double conditionedValue) {
-		return (caseValue.doubleValue() <= conditionedValue.doubleValue());
+		return (caseValue <= conditionedValue);
 	}
 
 	@Override
